update the methods and required fields according to feedback.

+++ src/main/java/tr/edu/ogu/ceng/dao/InternshipRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tr.edu.ogu.ceng.model.Internship;
 
-public interface InternshipRepository extends JpaRepository <Internship, Integer>{
+public interface InternshipRepository extends JpaRepository <Internship, Long>{
 	
 }
